only starts on start click

+++ client/components/game/enviroment.tsx
@@ -1,6 +1,9 @@
 import p5 from 'p5'
 import platform from './platform'
 import playerObject from './playerObject'
+import score from './score'
+import store from '../../store'
+// import store from '../../store'
 
 class enviroment {
   constructor(envSize) {
@@ -33,12 +36,19 @@ class enviroment {
     this.draw(envSize) //this will set up our canvas <--- and will setup our game loop <3
   }
 
+  startGame() {
+    return store.getState().game.start
+  }
+
   //this is where our game will take place
   update() {
+    this.startGame()
     this.drawScore()
-    this.createObjects()
-    this.deleteObjects()
-    this.panCamera()
+    if (this.startGame()) {
+      this.createObjects()
+      this.deleteObjects()
+      this.panCamera()
+    }
     this.player.playerInput(this.p5)
     this.platformArr.map((i) => {
       i.updatePlatform(this.p5, this.player)
@@ -50,7 +60,7 @@ class enviroment {
       this.height--
     }
     this.camera[1] = this.height
-    this.p5.translate(-this.camera[0], -this.camera[1])
+    this.p5.translate(-this.camera[0] + 100, -this.camera[1])
   }
   drawScore() {
     const c = [225, Math.random() * 50, 100]
